+++ QueryGenerator/src/com/miguelkvidal/querygenerator/api/condition/predicate/QPredicate.java
@@ -1,4 +1,4 @@
-package com.miguelkvidal.querygenerator.util.condition.predicate;
+package com.miguelkvidal.querygenerator.api.condition.predicate;
 
 import javax.persistence.criteria.Predicate;
