+++ QueryGenerator/src/com/miguelkvidal/querygenerator/api/expression/QAttribute.java
@@ -1,4 +1,4 @@
-package com.miguelkvidal.querygenerator.util.expression;
+package com.miguelkvidal.querygenerator.api.expression;
 
 import java.util.Map;
